refactor lint settings and SpecificIconType
