--- .eslintignore
@@ -1,1 +1,0 @@
-.eslintrc.js
