+++ src/example/components/basics/GearIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Icon as RNEIcon} from 'react-native-elements';
 
-import {SpecificIconType} from '../Constants';
+import {SpecificIconType} from './SpecificIconType';
 
 // React Native Elementsを使用したアイコンのサンプル
 const Icon = ({color = '#9D8CB8', ...props}: SpecificIconType) => {
