+++ src/example/components/basics/SpecificIconType.ts
